+++ roles/telegraf/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Missing dependency in filters
 
 ## [2.0.7] - 2020-10-20
 ### Added